Prefer using Doubles in Swift API and then converting to Float as needed

+++ AudioKit/Common/Nodes/Generators/Oscillators/FM Oscillator Bank/AKFMOscillatorBank.swift
@@ -291,8 +291,8 @@ open class AKFMOscillatorBank: AKPolyphonicNode, AKComponent {
     // MARK: - AKPolyphonic
     
     // Function to start, play, or activate the node at frequency
-    open override func play(noteNumber: MIDINoteNumber, velocity: MIDIVelocity, frequency:Float) {
-        internalAU?.startNote(noteNumber, velocity: velocity, frequency:frequency)
+    open override func play(noteNumber: MIDINoteNumber, velocity: MIDIVelocity, frequency: Double) {
+        internalAU?.startNote(noteNumber, velocity: velocity, frequency: Float(frequency))
     }
 
     /// Function to stop or bypass the node, both are equivalent
